refactor: Rename `lib/plugins/aws/custom-resources/generateZip.js`
