+++ lib/plugins/aws/custom-resources/generate-zip.js
