+++ lib/plugins/aws/custom-resources/index.js
@@ -6,7 +6,7 @@ const crypto = require('crypto');
 const BbPromise = require('bluebird');
 const fse = require('fs-extra');
 const { log } = require('@serverless/utils/log');
-const generateZip = require('./generateZip');
+const generateZip = require('./generate-zip');
 const ServerlessError = require('../../../serverless-error');
 
 const prepareCustomResourcePackage = _.memoize(async (zipFilePath) => {
